feat: refactor TheNav component and the form folder has been added
